requirements.txt
lxml
python-docx
pillow
